refactor: flatten gratuitous single-file directories
Collapse seven directories that each held exactly one file and carried
no taxonomy weight, moving the file up into its parent:

- src/controllers/progressView/backend/state/ProgressViewState.ts -> backend/
- src/auth/core/authCallback.ts -> src/auth/
- src/utils/prompt/index.ts -> src/utils/prompt.ts (import specifier
  @utils/prompt unchanged)
- packages/extension/src/progressView/frontend/contexts/streamContexts.ts
  and utils/audioNotification.ts -> frontend/ (removes the utils/ dir
  that shadowed the sibling utils.ts file)
- packages/extension/src/webview/frontend/contexts/mainViewContexts.ts
  and styles/fileSelectStyles.ts -> frontend/ (removes the styles/ dir
  that shadowed the sibling styles.ts file)

All import sites updated; no re-export shims left behind. Intentional
single-file directories are untouched: test-kernel mirrors of src/,
per-tool dirs under src/tools/, command-category dirs under
packages/extension/src/commands/, frozen @agent/* module layout, the
documented browser-safe @utils paths, and the cross-view backend/
split under src/controllers/.
