+++ packages/extension/src/progressView/frontend/audioNotification.ts
