+++ packages/extension/src/webview/frontend/components/LatexDiffsSection.ts
@@ -29,7 +29,7 @@ import type { TeXRAIconName } from '@shared/wa/iconNames';
 import { waIcon } from '@shared/wa/webAwesomeIcons';
 import { byString } from '@utils/core';
 import { MainViewEvents } from '../events';
-import { fileSelectLayoutStyles } from '../styles/fileSelectStyles';
+import { fileSelectLayoutStyles } from '../fileSelectStyles';
 
 type LatexDiffsAction = LatexDiffsActionDetail['action'];
