+++ packages/extension/src/webview/frontend/fileSelectStyles.ts
